backend/requirements.txt
fastapi==0.110.3
uvicorn[standard]==0.29.0
SQLAlchemy==2.0.29
psycopg2-binary==2.9.9
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
redis==5.0.4
pytest==8.1.1
httpx==0.27.0

